+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - New predefined constant `%_quail%`. It will point to the place where the top quail config catalog is.
+- Argument passing for commands via `quail -o ... command $ "arg"`
+- Argument passing for queues via `quail -o ... queue $ "arg1_c1, arg1_c2 arg2_c2, arg1_c3"`
 
 ### Changed